django_ai_app/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
openai>=1.14.0
python-dotenv>=1.0
